requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[cryptography]
PyJWT[Jwt]
pymysql
bcrypt
python-multipart
